Updated OUI parser to handle IEEE Registration Authority addresses
There's 3 cases where the IEEE Registration Authority has addresses...
one of them is for the first 3 octets, the other two are for addresses
that have a netmask of 36. In the two cases with a netmask of 36, the
first 3 octets are an entry before the list of netmasked addresses. Need
to ignore the initial entry or it will match addresses for all the
netmasked addresses.