Updated PCAP parser and Modbus protocol filter
PCAP parser now correctly checks to see if an IP interface already
exists for a given packet. If so, it upates the MAC address. If not, an
entire new chain of data from the node up to the IP interface is
created.

If the packet is detected as Modbus, the node is also tagged with
relevant Master/Slave data.